Refactor: make src dir, move scripts
