+++ __tests__/App-test.js
@@ -1,7 +1,7 @@
 import 'react-native';
 import React from 'react';
 import renderer from 'react-test-renderer';
-import App from '../App';
+import App from '../src/App';
 
 describe('App component test', () => {
   it('renders correctly', () => {
