+++ src/components/Search/styles.js
